# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Phase Snipe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trade history
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: executeTrade
      tags: [trades]
      summary: Execute a buy or sell trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
...
  /copy-trades:
    get:
      operationId: listCopyTrades
      tags: [copy-trades]
      summary: List all copy trade targets
      responses:
        "200":
          description: List of copy trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CopyTrade"
    post:
      operationId: createCopyTrade
      tags: [copy-trades]
      summary: Add a copy trade target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyTradeInput"
      responses:
        "201":
          description: Created copy trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrade"
  /copy-trades/{id}:
    get:
      operationId: getCopyTrade
      tags: [copy-trades]
      summary: Get copy trade by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CopyTrade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrade"
    patch:
      operationId: updateCopyTrade
      tags: [copy-trades]
      summary: Update copy trade target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyTradeUpdate"
      responses:
        "200":
          description: Updated copy trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrade"
    delete:
      operationId: deleteCopyTrade
      tags: [copy-trades]
      summary: Delete copy trade target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /copy-trades/{id}/start:
    post:
      operationId: startCopyTrade
      tags: [copy-trades]
      summary: Start copy trading
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrade"
  /copy-trades/{id}/pause:
    post:
      operationId: pauseCopyTrade
      tags: [copy-trades]
      summary: Pause copy trading
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrade"
  /limit-orders:
    get:
      operationId: listLimitOrders
      tags: [limit-orders]
      summary: List all limit orders
      responses:
        "200":
          description: List of limit orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LimitOrder"
    post:
      operationId: createLimitOrder
      tags: [limit-orders]
      summary: Create a limit order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitOrderInput"
      responses:
        "201":
          description: Created limit order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitOrder"
  /limit-orders/{id}:
    get:
      operationId: getLimitOrder
      tags: [limit-orders]
      summary: Get limit order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: LimitOrder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitOrder"
    patch:
      operationId: updateLimitOrder
      tags: [limit-orders]
      summary: Update limit order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitOrderUpdate"
      responses:
        "200":
          description: Updated limit order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitOrder"
    delete:
      operationId: deleteLimitOrder
      tags: [limit-orders]
      summary: Delete limit order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dca:
    get:
      operationId: listDcaSetups
      tags: [dca]
      summary: List all DCA setups
      responses:
        "200":
          description: List of DCA setups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DcaSetup"
    post:
      operationId: createDcaSetup
      tags: [dca]
      summary: Create a DCA setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DcaSetupInput"
      responses:
        "201":
          description: Created DCA setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaSetup"
  /dca/{id}:
    get:
      operationId: getDcaSetup
      tags: [dca]
      summary: Get DCA setup by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DcaSetup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaSetup"
    patch:
      operationId: updateDcaSetup
      tags: [dca]
      summary: Update DCA setup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DcaSetupUpdate"
      responses:
        "200":
          description: Updated DCA setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaSetup"
    delete:
      operationId: deleteDcaSetup
      tags: [dca]
      summary: Delete DCA setup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dca/{id}/start:
    post:
      operationId: startDcaSetup
      tags: [dca]
      summary: Start DCA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaSetup"
  /dca/{id}/pause:
    post:
      operationId: pauseDcaSetup
      tags: [dca]
      summary: Pause DCA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcaSetup"
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        type:
          type: string
          enum: [buy, sell]
        tokenSymbol:
          type: string
        tokenName:
          type: string
        contractAddress:
          type: string
        amountSol:
          type: number
        amountTokens:
          type: number
        priceSol:
          type: number
        pnlPercent:
          type: ["number", "null"]
        pnlSol:
          type: ["number", "null"]
        txHash:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, success, failed]
        executedAt:
          type: string
      required:
        - id
        - walletId
        - type
        - tokenSymbol
        - tokenName
        - contractAddress
        - amountSol
        - amountTokens
        - priceSol
        - txHash
        - status
        - executedAt
    TradeInput:
      type: object
      properties:
        type:
          type: string
          enum: [buy, sell]
        contractAddress:
          type: string
        amountSol:
          type: ["number", "null"]
        percentOfPosition:
          type: ["number", "null"]
        slippagePercent:
          type: number
        priorityFee:
          type: string
          enum: [auto, low, medium, high]
      required:
        - type
        - contractAddress
        - slippagePercent
        - priorityFee
...
    CopyTrade:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        targetAddress:
          type: string
        targetAlias:
          type: ["string", "null"]
        amountSol:
          type: number
        mode:
          type: string
          enum: [fixed, proportional]
        status:
          type: string
          enum: [active, paused, stopped]
        tradesCopied:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - walletId
        - targetAddress
        - amountSol
        - mode
        - status
        - tradesCopied
        - createdAt
    CopyTradeInput:
      type: object
      properties:
        targetAddress:
          type: string
        targetAlias:
          type: ["string", "null"]
        amountSol:
          type: number
        mode:
          type: string
          enum: [fixed, proportional]
      required:
        - targetAddress
        - amountSol
        - mode
    CopyTradeUpdate:
      type: object
      properties:
        targetAlias:
          type: ["string", "null"]
        amountSol:
          type: number
        mode:
          type: string
          enum: [fixed, proportional]
    LimitOrder:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        tokenSymbol:
          type: string
        contractAddress:
          type: string
        takeProfitPercent:
          type: ["number", "null"]
        stopLossPercent:
          type: ["number", "null"]
        trailingStopPercent:
          type: ["number", "null"]
        autoSell:
          type: boolean
        status:
          type: string
          enum: [active, triggered, cancelled]
        createdAt:
          type: string
      required:
        - id
        - walletId
        - tokenSymbol
        - contractAddress
        - autoSell
        - status
        - createdAt
    LimitOrderInput:
      type: object
      properties:
        contractAddress:
          type: string
        takeProfitPercent:
          type: ["number", "null"]
        stopLossPercent:
          type: ["number", "null"]
        trailingStopPercent:
          type: ["number", "null"]
        autoSell:
          type: boolean
      required:
        - contractAddress
        - autoSell
    LimitOrderUpdate:
      type: object
      properties:
        takeProfitPercent:
          type: ["number", "null"]
        stopLossPercent:
          type: ["number", "null"]
        trailingStopPercent:
          type: ["number", "null"]
        autoSell:
          type: boolean
    DcaSetup:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        tokenSymbol:
          type: string
        contractAddress:
          type: string
        amountSol:
          type: number
        intervalHours:
          type: number
        status:
          type: string
          enum: [active, paused, stopped]
        executionsCount:
          type: integer
        nextExecutionAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - walletId
        - tokenSymbol
        - contractAddress
        - amountSol
        - intervalHours
        - status
        - executionsCount
        - createdAt
    DcaSetupInput:
      type: object
      properties:
        contractAddress:
          type: string
        amountSol:
          type: number
        intervalHours:
          type: number
      required:
        - contractAddress
        - amountSol
        - intervalHours
    DcaSetupUpdate:
      type: object
      properties:
        amountSol:
          type: number
        intervalHours:
          type: number
...