# lib/api-zod/src/generated/types/dcaSetup.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phase Snipe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DcaSetupStatus } from './dcaSetupStatus';

export interface DcaSetup {
  id: number;
  walletId: number;
  tokenSymbol: string;
  contractAddress: string;
  amountSol: number;
  intervalHours: number;
  status: DcaSetupStatus;
  executionsCount: number;
  /** @nullable */
  nextExecutionAt?: string | null;
  createdAt: string;
}