# lib/api-zod/src/generated/types/dcaSetupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phase Snipe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DcaSetupInput {
  contractAddress: string;
  amountSol: number;
  intervalHours: number;
}