# lib/api-zod/src/generated/types/limitOrderUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Phase Snipe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LimitOrderUpdate {
  /** @nullable */
  takeProfitPercent?: number | null;
  /** @nullable */
  stopLossPercent?: number | null;
  /** @nullable */
  trailingStopPercent?: number | null;
  autoSell?: boolean;
}